Switch to better Let's Encrypt handling
This change require the newest version of core-base with the ssl-letsencrypt.sh script available with the TYPE parameter. For now we create a self-signed SSL certificate by default, after that we will try to find a cert by mdata and if that doesn't work we run Let's Encrypt.

With this change we switch to webroot authentication from Let's Encrypt, because a Standalone option is not working if a web server is already running.